requirements.txt
flask
flask-cors
langchain
langchain-community
langchain-chroma
langchain-core
langchain-groq
langchain-huggingface
langchain-text-splitters
pypdf
python-dotenv
sentence-transformers
